% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/As_svy_mean.R
\name{API_PAT}
\alias{API_PAT}
\title{API-PAT:Applied Pharmaceutical Innovation PK Analysis Tool}
\usage{
  API_PAT()
}
\description{
API-PAT: Applied Pharmaceutical Innovation PK Analysis Tool designed to perform Dose-Response and Non-Compartmental Analysis.
}

